Guessing Game Application
The application should receive two integers from the user, namely minimum and maximum and generate a random integer from minimum through maximum, inclusive. It then should give the user five chances to guess the integer. Each time the user makes a guess, the application should display one of two messages: "Guess higher" or "Guess lower". If the user guesses the generated number the application should let her/him know and also display the number of chances that were required for the user to guess the number. If the user is not able to guess the number in five attempts the application should display "Game over!: and indicate the correct number.