Credit Card Check Application
Credit card companies typically assign a special digit, called a check digit, to the end of each customer's credit card number. The check digit allows companies to verify that the credit card umber was entered accurately. The following method is used to verify actual credit card numbers but, for simplicity, we will describe it for numbers with 8 digits instead of 16.